+++ .vscode/settings.json
@@ -12,7 +12,9 @@
         "stdlib.h": "c",
         "string.h": "c",
         "time.h": "c",
-        "math.h": "c"
+        "math.h": "c",
+        "stat.h": "c",
+        "file.h": "c"
     },
     "C_Cpp.errorSquiggles": "disabled"
 }
