+++ lib/array.c
