+++ lib/key.c
@@ -1,0 +1,12 @@
+// Key
+#define EXIST_KEY 27
+#define ENTER_KEY 13
+#define DELETE_KEY 83
+#define BACKSPACE 8
+
+// Arrow Key Code
+#define UP_KEY 72
+#define DOWN_KEY 80
+#define LEFT_KEY 75
+#define RIGHT_KEY 77
+
